More validations plus changes to json files required after the move from camelCase to snake_case.

+++ data-claims-event-service/src/test/java/uk/gov/justice/laa/dstew/payments/claimsevent/mapper/FeeSchemeMapperTest.java
@@ -33,7 +33,7 @@ class FeeSchemeMapperTest {
               .disbursementsVatAmount(BigDecimal.valueOf(1.03))
               .isVatApplicable(true)
               .priorAuthorityReference("disbursementPriorAuthority")
-              .adjournedHearingFeeAmount(BigDecimal.valueOf(1))
+              .adjournedHearingFeeAmount(1)
               .cmrhOralCount(2)
               .cmrhTelephoneCount(3)
               // TODO: add additional travel property
@@ -100,7 +100,7 @@ class FeeSchemeMapperTest {
 
       ClaimResponse claim =
           new ClaimResponse()
-              .adjournedHearingFeeAmount(BigDecimal.valueOf(1))
+              .adjournedHearingFeeAmount(1)
               .cmrhOralCount(2)
               .cmrhTelephoneCount(3)
               // TODO: add additional travel property
